project struct refactoring
