+++ notification/src/main/java/com/chronica/notification/data/constant/SortDirection.java
@@ -6,13 +6,13 @@ import org.springframework.data.domain.Sort;
 
 @Getter
 public enum SortDirection {
-    ASC("ASC", Sort.Direction.ASC),
-    DESC("DESC", Sort.Direction.DESC);
+    ASC(Sort.Direction.ASC),
+    DESC(Sort.Direction.DESC);
 
-    private final String name;
+
     private final Sort.Direction dir;
-    SortDirection(String name, Sort.Direction dir) {
-        this.name = name;
+    SortDirection(Sort.Direction dir) {
+
         this.dir = dir;
     }
 }
