+++ notification/src/main/java/com/chronica/notification/data/mapper/injection/AlertMapper.java
@@ -1,4 +1,4 @@
-package com.chronica.notification.data.mapper;
+package com.chronica.notification.data.mapper.injection;
 
 import com.chronica.notification.data.dto.AlertDTO;
 import com.chronica.notification.data.dto.integrant.NotificationDTO;
