+++ notification/src/main/java/com/chronica/notification/data/mapper/injection/MapperConfig.java
@@ -2,7 +2,7 @@ package com.chronica.notification.data.mapper.injection;
 
 import com.chronica.notification.data.dto.abstraction.BaseDTO;
 import com.chronica.notification.data.entity.Notification;
-import com.chronica.notification.data.mapper.implementation.MapperModelImpl;
+import com.chronica.notification.data.mapper.MapperMajorNotification;
 import com.chronica.notification.data.mapper.implementation.MapperImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,8 +11,8 @@ import org.springframework.context.annotation.Configuration;
 public class MapperConfig {
 
     @Bean
-    public MapperModelImpl<BaseDTO, Notification> mapperConfiguration(){
-        return new MapperModelImpl<>(new MapperImpl<>());
+    public MapperMajorNotification<BaseDTO, Notification> mapperConfiguration(){
+        return new MapperMajorNotification<>(new MapperImpl<>());
     }
