+++ notification/src/main/java/com/chronica/notification/logic/notification/NotificationServiceConfig.java
@@ -3,11 +3,6 @@ package com.chronica.notification.logic.notification;
 import com.chronica.notification.data.entity.Alert;
 import com.chronica.notification.data.entity.Invitation;
 import com.chronica.notification.data.entity.Message;
-import com.chronica.notification.data.mapper.AlertMapper;
-import com.chronica.notification.data.mapper.InvitationMapper;
-import com.chronica.notification.data.mapper.MessageMapper;
-import com.chronica.notification.data.mapper.implementation.MapperImpl;
-import com.chronica.notification.data.mapper.injection.NotificationMapper;
 import com.chronica.notification.data.repository.AlertRepository;
 import com.chronica.notification.data.repository.InvitationRepository;
 import com.chronica.notification.data.repository.MessageRepository;
